Move element operations headers to folder matching namespace
