+++ core/include/scipp/core/element/event_operations.h
