+++ core/include/scipp/core/element/geometric_operations.h
