+++ core/include/scipp/core/element/trigonometry_operations.h
