+++ core/include/scipp/core/element/unary_operations.h
