+++ core/test/element_event_operations_test.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2020 Scipp contributors (https://github.com/scipp)
 #include <gtest/gtest.h>
 
-#include "scipp/core/element_event_operations.h"
+#include "scipp/core/element/event_operations.h"
 #include "scipp/core/values_and_variances.h"
 
 #include "fix_typed_test_suite_warnings.h"
