+++ core/test/element_trigonometry_operations_test.cpp
@@ -3,7 +3,7 @@
 #include <gtest/gtest.h>
 
 #include "scipp/common/constants.h"
-#include "scipp/core/element_trigonometry_operations.h"
+#include "scipp/core/element/trigonometry_operations.h"
 #include "scipp/units/unit.h"
 
 #include "fix_typed_test_suite_warnings.h"
