+++ variable/test/operations_test.cpp
@@ -6,7 +6,7 @@
 #include "test_macros.h"
 #include "fix_typed_test_suite_warnings.h"
 
-#include "scipp/core/element_unary_operations.h"
+#include "scipp/core/element/unary_operations.h"
 #include "scipp/variable/except.h"
 #include "scipp/variable/operations.h"
 #include "scipp/variable/variable.h"
