+++ variable/trigonometry.cpp
@@ -5,7 +5,7 @@
 #include <cmath>
 
 #include "scipp/common/constants.h"
-#include "scipp/core/element_trigonometry_operations.h"
+#include "scipp/core/element/trigonometry_operations.h"
 #include "scipp/variable/transform.h"
 #include "scipp/variable/trigonometry.h"
